Reworked Backend and frontend
App can now save task items, and display them again. the view for this moves in from the bottom of the screen, the camera now succesfully saves images taken with the camera, and can also display them again. and finally a list item can be checked off (not rechecked again, but come on its still in development)